Add cython version of counting

+++ tests/test_counting.py
@@ -3,11 +3,11 @@ import numpy as np
 from numpy.testing import assert_equal
 
 def test_count_no_object():
-    a = np.zeros([6, 6], dtype=np.float64)
+    a = np.zeros([6, 6], dtype=np.int32)
     assert_equal(count_objects(a), 0)
 
 def test_count_1_object():
-    a = np.zeros([6, 6], dtype=np.float64)
+    a = np.zeros([6, 6], dtype=np.int32)
     a[1:-1, 1:-1] = 1
     assert_equal(count_objects(a), 1)
 
@@ -22,7 +22,7 @@ def test_count_objects_2_objects():
        [ 0.,  0.,  0.,  0.,  0.,  0.,  0.,  0.,  0.,  0.],
        [ 0.,  0.,  0.,  0.,  0.,  0.,  0.,  0.,  0.,  0.],
        [ 0.,  0.,  0.,  0.,  0.,  0.,  0.,  0.,  0.,  0.],
-       [ 0.,  0.,  0.,  0.,  0.,  0.,  0.,  0.,  0.,  0.]])
+       [ 0.,  0.,  0.,  0.,  0.,  0.,  0.,  0.,  0.,  0.]], dtype=np.int32)
     assert_equal(count_objects(a), 2)
 
 def test_count_1_object_rectangle():
@@ -36,5 +36,5 @@ def test_count_1_object_rectangle():
        [ 0.,  0.,  0.,  0.,  0.],
        [ 0.,  0.,  0.,  0.,  0.],
        [ 0.,  0.,  0.,  0.,  0.],
-       [ 0.,  0.,  0.,  0.,  0.]])
+       [ 0.,  0.,  0.,  0.,  0.]], dtype=np.int32)
     assert_equal(count_objects(a), 1)
